% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_crossref.R
\name{fetch_crossref_messages}
\alias{fetch_crossref_messages}
\title{Fetch Crossref records by DOI}
\usage{
fetch_crossref_messages(dois)
}
\arguments{
\item{dois}{Character vector of DOIs.}
}
\value{
List of Crossref \code{message} objects.
}
\description{
Fetch Crossref records by DOI
}
