% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_crossref.R
\name{fetch_crossref_works}
\alias{fetch_crossref_works}
\title{Fetch Crossref works}
\usage{
fetch_crossref_works(filters = NULL, query = NULL, limit = 100L)
}
\arguments{
\item{filters}{Named list of Crossref filters.}

\item{query}{Optional free-text query.}

\item{limit}{Maximum number of items to return.}
}
\value{
List of Crossref items.
}
\description{
Fetch Crossref works
}
