% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_schema.R
\name{litxr_anchor_references_template}
\alias{litxr_anchor_references_template}
\title{Template row for anchor references}
\usage{
litxr_anchor_references_template()
}
\value{
One-row \code{data.table} with the anchor references schema.
}
\description{
Template row for anchor references
}
