% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_bm25.R
\name{litxr_bm25_load_index}
\alias{litxr_bm25_load_index}
\title{Load a BM25 index}
\usage{
litxr_bm25_load_index(index_dir)
}
\arguments{
\item{index_dir}{Directory containing BM25 index files.}
}
\value{
List with \code{docs}, \code{vocab}, \code{postings}, and \code{meta}.
}
\description{
Load a BM25 index
}
