% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_bm25.R
\name{litxr_bm25_search}
\alias{litxr_bm25_search}
\title{Search a BM25 index}
\usage{
litxr_bm25_search(
  query,
  index,
  top_k = 50L,
  k1 = 1.5,
  b = 0.75,
  restrict_doc_ids = NULL,
  return_terms = TRUE
)
}
\arguments{
\item{query}{Free-text query string.}

\item{index}{Loaded BM25 index from \code{litxr_bm25_load_index()}.}

\item{top_k}{Number of ranked rows to return.}

\item{k1}{BM25 \code{k1} parameter.}

\item{b}{BM25 \code{b} parameter.}

\item{restrict_doc_ids}{Optional character vector restricting candidate docs.}

\item{return_terms}{Whether to include \code{matched_terms}.}
}
\value{
\code{data.table} of ranked BM25 results.
}
\description{
Search a BM25 index
}
