% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_bm25.R
\name{litxr_bm25_search_query_set}
\alias{litxr_bm25_search_query_set}
\title{Search a BM25 index from one lexical query set}
\usage{
litxr_bm25_search_query_set(
  query_sets,
  query_set,
  category = NULL,
  index,
  top_k = 50L,
  k1 = 1.5,
  b = 0.75,
  restrict_doc_ids = NULL
)
}
\arguments{
\item{query_sets}{Named nested query-set list.}

\item{query_set}{Query-set id to use.}

\item{category}{Optional category id inside the query set.}

\item{index}{Loaded BM25 index.}

\item{top_k}{Number of ranked rows to return.}

\item{k1}{BM25 \code{k1} parameter.}

\item{b}{BM25 \code{b} parameter.}

\item{restrict_doc_ids}{Optional candidate doc restriction.}
}
\value{
\code{data.table} of ranked BM25 results.
}
\description{
Search a BM25 index from one lexical query set
}
