% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_schema.R
\name{litxr_citation_logic_nodes_template}
\alias{litxr_citation_logic_nodes_template}
\title{Template row for citation logic nodes}
\usage{
litxr_citation_logic_nodes_template()
}
\value{
One-row \code{data.table} with the citation logic nodes schema.
}
\description{
Template row for citation logic nodes
}
