% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{litxr_config_path}
\alias{litxr_config_path}
\title{Return the litxr config path}
\usage{
litxr_config_path()
}
\value{
Character scalar path.
}
\description{
\code{litxr} resolves the config file as \code{file.path(LITXR_DATA_ROOT, "config.yaml")}.
}
