% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_schema.R
\name{litxr_descriptive_stats_template}
\alias{litxr_descriptive_stats_template}
\title{Template row for descriptive statistics}
\usage{
litxr_descriptive_stats_template()
}
\value{
One-row \code{data.table} with the descriptive statistics schema.
}
\description{
Template row for descriptive statistics
}
