% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_search.R
\name{litxr_label_collection_by_category}
\alias{litxr_label_collection_by_category}
\title{Label one collection by category score threshold}
\usage{
litxr_label_collection_by_category(
  collection_id,
  query_set_id,
  config = NULL,
  field = "abstract",
  model,
  threshold = 0.5,
  aggregations = c("max"),
  top_k = 3L,
  ref_ids = NULL,
  date_from = NULL,
  date_to = NULL,
  chunk_size = 256L
)
}
\description{
Label one collection by category score threshold
}
