% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_category_label.R
\name{litxr_lexical_escape_regex}
\alias{litxr_lexical_escape_regex}
\title{Escape regex metacharacters for lexical matching}
\usage{
litxr_lexical_escape_regex(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Escaped character vector.
}
\description{
Escape regex metacharacters for lexical matching
}
