% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_category_label.R
\name{litxr_lexical_keyword_pattern}
\alias{litxr_lexical_keyword_pattern}
\title{Build a keyword regex pattern}
\usage{
litxr_lexical_keyword_pattern(
  keyword,
  case_insensitive = TRUE,
  whole_term = TRUE
)
}
\arguments{
\item{keyword}{Keyword or phrase.}

\item{case_insensitive}{Whether the pattern should be case-insensitive.}

\item{whole_term}{Whether to require Unicode word boundaries.}
}
\value{
Regex pattern string.
}
\description{
Build a keyword regex pattern
}
