% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_common.R
\name{litxr_lexical_tokenize}
\alias{litxr_lexical_tokenize}
\title{Tokenize lexical text}
\usage{
litxr_lexical_tokenize(x, normalize = TRUE)
}
\arguments{
\item{x}{Character vector.}
}
\value{
List of character vectors.
}
\description{
Tokenizes normalized text into Unicode-aware term vectors.
}
