% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{litxr_list_collections}
\alias{litxr_list_collections}
\title{List collections registered in the config}
\usage{
litxr_list_collections(config = NULL)
}
\arguments{
\item{config}{Optional parsed config list or a direct config path. When
omitted, \code{litxr} reads \code{LITXR_CONFIG}.}
}
\value{
\code{data.table} of collection registrations.
}
\description{
List collections registered in the config
}
