% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_schema.R
\name{litxr_paper_type_levels}
\alias{litxr_paper_type_levels}
\title{Canonical paper type levels}
\usage{
litxr_paper_type_levels()
}
\value{
Character vector of canonical paper-type levels.
}
\description{
Canonical paper type levels
}
