% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{litxr_read_config}
\alias{litxr_read_config}
\title{Read a litxr config file}
\usage{
litxr_read_config(path = ".")
}
\arguments{
\item{path}{Project root or direct path to a config file.}
}
\value{
Parsed config list.
}
\description{
Read a litxr config file
}
