% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{litxr_read_journal}
\alias{litxr_read_journal}
\title{Read locally stored records for one journal}
\usage{
litxr_read_journal(journal_id, config = ".")
}
\arguments{
\item{journal_id}{Journal identifier from \code{config.yaml}.}

\item{config}{Parsed config list or a path that \code{litxr_read_config()} accepts.}
}
\value{
\code{data.table} of local records.
}
\description{
Read locally stored records for one journal
}
