% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_schema.R
\name{litxr_standardized_findings_template}
\alias{litxr_standardized_findings_template}
\title{Template row for standardized findings}
\usage{
litxr_standardized_findings_template()
}
\value{
One-row \code{data.table} with the standardized findings schema.
}
\description{
Template row for standardized findings
}
