% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{litxr_sync_all}
\alias{litxr_sync_all}
\title{Sync all configured journals}
\usage{
litxr_sync_all(config = ".")
}
\arguments{
\item{config}{Parsed config list or a path that \code{litxr_read_config()} accepts.}
}
\value{
Named list of synced \code{data.table}s.
}
\description{
Sync all configured journals
}
