% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_schema.R
\name{litxr_validate_anchor_references}
\alias{litxr_validate_anchor_references}
\title{Validate anchor references rows}
\usage{
litxr_validate_anchor_references(anchors)
}
\arguments{
\item{anchors}{Data frame, data.table, or named list of anchor-reference
rows.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Validate anchor references rows
}
