% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_arxiv.R
\name{parse_arxiv_entry_unified}
\alias{parse_arxiv_entry_unified}
\title{Parse one arXiv entry into the unified litxr schema}
\usage{
parse_arxiv_entry_unified(entry)
}
\arguments{
\item{entry}{One arXiv Atom entry node.}
}
\value{
One-row \code{data.table}.
}
\description{
Parse one arXiv entry into the unified litxr schema
}
