% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_crossref.R
\name{parse_crossref_entry_unified}
\alias{parse_crossref_entry_unified}
\title{Parse one Crossref work into the unified litxr schema}
\usage{
parse_crossref_entry_unified(cr_message)
}
\arguments{
\item{cr_message}{One Crossref \code{message} object.}
}
\value{
One-row \code{data.table}.
}
\description{
Parse one Crossref work into the unified litxr schema
}
