% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_arxiv.R
\name{xml_text_or_na}
\alias{xml_text_or_na}
\title{Extract text from an XML node or return \code{NA}}
\usage{
xml_text_or_na(node, xpath, ns = NULL)
}
\arguments{
\item{node}{XML node.}

\item{xpath}{XPath expression.}

\item{ns}{Optional XML namespace mapping.}
}
\value{
Character scalar.
}
\description{
Extract text from an XML node or return \code{NA}
}
