% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_arxiv.R
\name{fetch_arxiv_xml}
\alias{fetch_arxiv_xml}
\title{Fetch arXiv API XML}
\usage{
fetch_arxiv_xml(
  id_vec = NULL,
  search_query = NULL,
  start = NULL,
  max_results = 100L,
  retry_max = 3L
)
}
\arguments{
\item{id_vec}{Optional arXiv ids.}

\item{search_query}{Optional arXiv search query.}

\item{start}{Result offset.}

\item{max_results}{Maximum number of records to request.}

\item{retry_max}{Maximum number of retries after rate limiting or transient
request failure.}
}
\value{
XML document.
}
\description{
Fetch arXiv API XML
}
