% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{litxr_config_path}
\alias{litxr_config_path}
\title{Return the litxr config path from \code{LITXR_CONFIG}}
\usage{
litxr_config_path()
}
\value{
Character scalar path.
}
\description{
\code{litxr} reads the config path from the \code{LITXR_CONFIG} environment variable.
Users should set that variable in \code{.Renviron}.
}
