% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{litxr_config_path}
\alias{litxr_config_path}
\title{Return the default litxr config path}
\usage{
litxr_config_path(path = ".")
}
\arguments{
\item{path}{Project root.}
}
\value{
Character scalar path.
}
\description{
\code{litxr} keeps the user-specific project config in \code{.litxr/config.yaml} at the
project root so the package can keep local sync settings out of version
control.
}
