% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{litxr_list_journals}
\alias{litxr_list_journals}
\title{List journals registered in the config}
\usage{
litxr_list_journals(config = ".")
}
\arguments{
\item{config}{Parsed config list or a path that \code{litxr_read_config()} accepts.}
}
\value{
\code{data.table} of journal registrations.
}
\description{
List journals registered in the config
}
