% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{litxr_llm_digest_template}
\alias{litxr_llm_digest_template}
\title{Create a default structured LLM digest template}
\usage{
litxr_llm_digest_template(ref_id, schema_version = "v2")
}
\arguments{
\item{ref_id}{Reference identifier.}

\item{schema_version}{Digest schema version. Supported values are \code{"v2"}
and \code{"v3"}.}
}
\value{
Named list representing the default digest schema.
}
\description{
Create a default structured LLM digest template
}
