% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{litxr_read_config}
\alias{litxr_read_config}
\title{Read a litxr config file}
\usage{
litxr_read_config(path = NULL)
}
\arguments{
\item{path}{Optional direct path to a config file. When omitted, \code{litxr}
reads \code{LITXR_CONFIG}.}
}
\value{
Parsed config list.
}
\description{
Read a litxr config file
}
