% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{litxr_sync_all}
\alias{litxr_sync_all}
\title{Sync all configured collections}
\usage{
litxr_sync_all(config = NULL)
}
\arguments{
\item{config}{Optional parsed config list or a direct config path. When
omitted, \code{litxr} reads \code{LITXR_CONFIG}.}
}
\value{
Named list of synced \code{data.table}s.
}
\description{
Sync all configured collections
}
